STACK-POS-7: set MCU params and remote mcu params update in RemoteMCUControllerRepository

+++ NxGnFirebaseModule/src/main/java/com/vismo/nxgnfirebasemodule/model/MeterSdkConfiguration.kt
@@ -19,7 +19,7 @@ data class OperatingAreaConfig(
 )
 
 data class DashFeeItem(
-    @SerializedName("dash_fee_constant") val dashFeeConstant: Int,
+    @SerializedName("dash_fee_constant") val dashFeeConstant: Double,
     @SerializedName("dash_fee_rate") val dashFeeRate: Double,
     @SerializedName("operating_area") val operatingArea: OperatingArea
 )
